refactored to include train/classify split; added argument flags
I refactored main() into train and annotate (or classify) to allow
for simple training loops vs. annotation or classification loops.
The model is now saved via pickle to the model directory to be
loaded later.  Annotated notes are not yet written to disk so there
is no way to score them.  (Scoring will need to be done via ETUDE
most likely although an initial scoring can be done internal to
the SVM modeler but that's not a good long-term solution.)