added Sentence and NoteSection types; load from typesDir
The types-dir is a new command line argument where we can start
collecting XML definitions of types (to make for cleaner code)

The default output for sentences was upgraded to the SHARPn-style
Sentence annotation type.  The default header type was converted
to NoteSection (rather than a binary flag on the old TBD:sentence
type).